move utility functions to pvutils.cpp
